Added Graph Tab
Added a graph tab so that the page can switch over to showing the graph over time of pollutants. Had chatGPT generate some data points and saved that as data2.csv . Currently getting a SettingWithCopyWarning whenever I generate a new graph, but it doesn't seem to affect making the graph?